+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add a dialog for editing an existing item.
+
 ## [0.1.0] - 2026-03-21
 
 ### Added